Made various test updates and seperated out
