+++ .github/workflows/python_tests.yml
@@ -4,6 +4,7 @@ on:
   workflow_dispatch:
   push:
     branches:
+      - feat/backend
       - dev
   pull_request:
     branches:
@@ -12,7 +13,7 @@ on:
 env:
     S3_BUCKET_NAME: ${{ secrets.S3_BUCKET_NAME }}
 jobs:
-  test:
+  End_to_End_Tests_Route_Testing:
     runs-on: ubuntu-latest
 
     steps:
@@ -53,6 +54,32 @@ jobs:
 
     - name: Run S3 Connection Test
       env:
-        S3_BUCKET_NAME: ${{ secrets.S3_BUCKET_NAME }}  # Pass the S3_BUCKET_NAME secret
+        S3_BUCKET_NAME: ${{ secrets.S3_BUCKET_NAME }}
         S3_FILE_KEY: ${{ secrets.S3_FILE_KEY }}
-      run: python -m pytest tests/unit_testing.py
+      run: python -m pytest tests/test_s3_bucket_connection.py
+
+  Test_file_exists:
+        runs-on: ubuntu-latest
+    
+        steps:
+        - name: Checkout code
+          uses: actions/checkout@v3
+    
+        - name: Configure AWS credentials
+          uses: aws-actions/configure-aws-credentials@v3
+          with:
+            aws-access-key-id: ${{ secrets.AWS_ACCESS_KEY_ID }}
+            aws-secret-access-key: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
+            aws-region: eu-west-1
+    
+        - name: Install dependencies
+          run: |
+            python -m pip install --upgrade pip
+            pip install -r requirements.txt
+    
+        - name: Run S3 Connection Test
+          env:
+            S3_BUCKET_NAME: ${{ secrets.S3_BUCKET_NAME }}
+            S3_FILE_KEY: ${{ secrets.S3_FILE_KEY }}
+          run: python -m pytest tests/test_s3_file_exists.py
+
